Draw and Erase obstacles with mouse. Renamed "InitializePathfinding" to "PathfindingManager".
