+++ Assets/Gfx/Materials/Cursor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 061ae88b568e6b74eaba8eeb7ca3973c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
