+++ Assets/Scripts/TextFade.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e73e26decf75f5b49b06fba90116801d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
